sistema/backend/requirements.txt
djangorestframework==3.8.2
django-filter==2.0.0
django==2.1.1
msgpack==0.5.6
django-cors-headers==2.4.0